Package picos_std

This package contains sample scheduler agnostic libraries for Picos. Many of the modules are intentionally designed to mimic modules from the OCaml Stdlib.

Picos_std_finally Syntax for avoiding resource leaks for Picos.

Picos_std_event Basic event abstraction for Picos.

Picos_std_structured Basic structured concurrency primitives for Picos.

Picos_std_sync Basic communication and synchronization primitives for Picos.

These libraries are both meant to serve as examples of what can be done and to also provide practical means for programming with fibers. Hopefully there will be many more libraries implemented in Picos like these providing different approaches, patterns, and idioms for structuring concurrent programs.

Package info

changes-files

CHANGES.md

license-files

LICENSE.md

readme-files

README.md